% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{A EM algorithm for Gaussian mixture models}
\usage{
EM(x, mu, sigma1, sigma2, prop)
}
\arguments{
\item{x}{e}

\item{mu}{initial mean}

\item{sigma1}{initial sigma1}

\item{sigma2}{initial sigma2}

\item{prop}{initial proportion}
}
\value{
estimates
}
\description{
A EM algorithm for Gaussian mixture models
}
\examples{
\dontrun{
mu<-matrix(c(2,4,50,80),2,2)
sigma1<-matrix(c(1,0,0,1),2,2)
sigma2<-matrix(c(1,0,0,1),2,2)
prop<-c(0.5,2)
EM(data,mu,sigma1,sigma2,prop)
}
}
