% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLMEV.R
\name{PLMEV.rob.2}
\alias{PLMEV.rob.2}
\title{PLMEV procedure}
\usage{
PLMEV.rob.2(y, vv, xt, palabra = "mcd", cw = 0.975, k = 0.55, ventana)
}
\arguments{
\item{y}{observed response Y}

\item{vv}{observed data V}

\item{xt}{observed t in nonparametric components}

\item{palabra}{selects the robust estimators of multivariate location and scatter, "mcd": MCD- (minimum covariance determinant) estimators, "m": M- estimators, "s": S- estimators, default is "mcd"}

\item{cw}{quantile of a chi^2, default is 0.975}

\item{k}{tuning constant of score function, default 0.55 (90% efficiency under normality)}

\item{ventana}{smoothing parameter}
}
\value{
estimates
}
\description{
a robust estimating method in semiparametric partially linear errors-in-variables model
}
\examples{
\dontrun{
n=200
z1 <- rnorm(n)
z2 <- rnorm(n)
eps <- rnorm(n,0,0.2)
eps1 <- rnorm(n,0,0.2)
eps2 <- rnorm(n,0,0.2)
xx <- cbind(z1,z2)
xt <- runif(n,-6,6)
vv <- cbind(z1+eps1,z2+eps2)
y <- 2*xx[,1] + xx[,2] + cos(pi*xt/6) + eps
gg <- cos(pi*xt/6)
ventana=2.5
salida=PLMEV.rob.2(y,vv,xt,palabra="mcd",cw=0.975,k=0.55,ventana)
}
}
