\name{SC19062-package}
\alias{SC19062-package}
\alias{SC19062}
\docType{package}
\title{
  Two functions with homework answers
}
\description{
  This package includes two main functions. One is EM alogithm for Gaussian mixture models. The other is a robust estimating method in semiparametric partially linear errors-in    -variables models. All homework answers for 'Statistical computing' course are following.
}
\author{
Zheng Zhi

Maintainer: Zheng Zhi <zheng017@mail.ustc.edu.cn>
}
\keyword{ package }

